requirements.txt
# GIS парсинг
fiona>=1.9
geopandas>=0.14
pyproj>=3.6
shapely>=2.0

# MCP Server (Фаза 2)
fastmcp>=0.1

# Ingestion API (Фаза 3)
fastapi>=0.111
uvicorn[standard]>=0.29
python-multipart>=0.0.9
python-jose[cryptography]>=3.3   # JWT

# 7z для .atbx (опционально, Фаза 1.3)
py7zr>=0.21

# Утилиты
python-dotenv>=1.0
numpy>=1.26.0
requests
jinja2